# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node.js Project Demo",
    "version": "1.0.0",
    "description": "Node.js Project Demo APIs documentation",
    "licence": {
      "name": "Asif Sk"
    }
  },
  "host": "localhost:3001",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "in": "cookie",
      "name": "token"
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "paths": {
...
    "/posts/all-post": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get all posts when user is Logged-In",
        "responses": {
          "200": {
            "description": "Details of Post",
            "schema": {
              "$ref": "#/definitions/allposts"
            }
          },
          "401": {
            "description": "Unauthorized Access"
          }
        }
      }
    },
...
    "/posts/my-post": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get all post of a particular user",
        "responses": {
          "200": {
            "description": "Comment deleted from the post"
          },
          "401": {
            "description": "Unauthorized Access"
          },
          "404": {
            "description": "Either comment not found or post not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "allposts": {
      "required": [
        "postTitle",
        "postImage",
        "postContent"
      ],
      "properties": {
        "postTitle": {
          "type": "string"
        },
        "postImage": {
          "type": "string"
        },
        "postContent": {
          "type": "string"
        }
      }
    },
...
  }
}